+++ CHANGELOG.md
@@ -6,9 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## Unreleased
-/
 
-## 2.0.x
+## 2.1.0 - 2022.02.01
+### Added
+- `/docs` and [RTFD](https://AlbertUnruhUtilsPy.rtfd.io) is online! (Docs for old versions 'll follow soon (and won't be mentioned here))
+- `.utils.decorator` (simple decorator you can use over time)
+- `.visual.tex` (to create tex-images from str)
+
+### Fixed
+- added dates to versions
+
+## 2.0.0 - 2022.01.30
 ### Added
 - CHANGELOG.md was finally added
 - README.md shows instructions how to install the `.asynchronous.*` part of the package
@@ -20,7 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - sorry PEP8, now you are more present (`__all__` is now at the top of the file and imports are alphabetically sorted)
 
-## 1.0.x
+## 1.0.0 - 2022.01.29
 ### Added
 - `.utils.logger` (for logging)
 - `.utils.version` (to compare versions)